fix: swiping feature changes for the commentary part -now swiping in the commentary tab will switch to the next or previous chapter and stay on the selected explanation
