+++ packages/frontend-base/src/hooks/useSearchParams.ts
@@ -50,6 +50,8 @@ export const useSaveSearchParams = () => {
   return { saveSearchParams, saveBibleVersionOnURL };
 };
 
+import { fetchAllChaptersByBook } from "./useBible";
+
 export const useGetSearchParams = () => {
   const searchParams = useSearchParams();
   const bookId = Number(searchParams?.get("bookId")) || 1;
@@ -63,6 +65,8 @@ export const useGetSearchParams = () => {
     (searchParams?.get("explanationType") as ExplanationTypeEnum) ||
     ExplanationTypeEnum.summary;
 
+  const { chapters } = fetchAllChaptersByBook(bookId);
+
   return {
     bookId,
     verseId,
@@ -71,5 +75,6 @@ export const useGetSearchParams = () => {
     conversationId,
     explanationId,
     explanationType,
+    chapters,
   };
 };
